fix: Label Browser Button #220
