+++ src/components/LabelBrowser/components/LabelsSelector/LabelsList.js
@@ -1,20 +1,7 @@
 import { useMemo } from "react";
 import styled from "@emotion/styled";
-import ShowLogsButton from "../Buttons/ShowLogsButton";
-import { queryBuilder } from "../../helpers/querybuilder";
-import { useDispatch, useSelector } from "react-redux";
-import { setLeftPanel } from "../../../../actions/setLeftPanel";
-import { setRightPanel } from "../../../../actions/setRightPanel";
-import store from "../../../../store/store";
 
-function panelAction(name, value) {
-    if (name === "left") {
-        return setLeftPanel(value);
-    }
-    return setRightPanel(value);
-}
 export const EmptyLabels = (props) => {
-    
     const EmptyCont = styled.div`
         color: ${({ theme }) => theme.textColor};
         padding: 10px;
@@ -24,7 +11,6 @@ export const EmptyLabels = (props) => {
 };
 
 // split labels for metrics and labels for logs
-
 
 function LabelItem(props) {
     const { selected, label, type } = props;
@@ -51,10 +37,8 @@ function LabelItem(props) {
 }
 
 export default function LabelsList(props) {
-    const dispatch = useDispatch();
-    const { labels, data, name } = props;
+    const { labels, data } = props;
     const { dataSourceType } = data;
-    const panelQuery = useSelector((store) => store[name]);
 
     const onClick = (e) => {
         if (e === "Select Metric") {
@@ -62,11 +46,6 @@ export default function LabelsList(props) {
         }
         props.onLabelSelected(e);
     };
-
-    const onMetricOptionsClick = (e) => {
-
-        console.log(e)
-    }
 
     const lsList = useMemo(() => {
         if (dataSourceType !== "metrics") {
@@ -76,31 +55,9 @@ export default function LabelsList(props) {
         return labels?.filter((f) => f.name !== "__name__");
     }, [dataSourceType, labels]);
 
-    const useQuery = () => {
-        const qu = queryBuilder(data.labels, data.expr)
-        const panel = [...panelQuery];
-        panel.forEach((query) => {
-            if (query.id === props.data.id) {
-                query.expr = qu;
-                query.labels = data.labels
-            }
-        });
-        dispatch(panelAction(name, panel));
-    }
     return (
         <div className="valuelist-content">
-            {/* {metricLabel !== null && lsList && dataSourceType === 'metrics' && (
-                <LabelItem
-                    type={"metric"}
-                 
-                    label={"Select Metric"}
-                    selected={metricLabel?.selected}
-                    onClick={onMetricOptionsClick}
-                 //   onClick={onClick}
-                />
-            )} */}
-
-            {lsList &&
+            {lsList && (
                 <>
                     {lsList.map((label, key) => (
                         <LabelItem
@@ -110,15 +67,9 @@ export default function LabelsList(props) {
                             onClick={onClick}
                         />
                     ))}
-                    <ShowLogsButton
-                        onClick={useQuery}
-                        isMobile={false}
-                        alterText={"Use Query"}
-                        loading={false}
-                    />
                 </>
-            }
-                
+            )}
+
             {!lsList && <EmptyLabels {...props} />}
         </div>
     );
