+++ src/components/LabelBrowser/components/LabelsSelector/ValuesSelector.js
@@ -1,11 +1,23 @@
 import { useEffect, useMemo, useState } from "react";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
+import { setLeftPanel } from "../../../../actions/setLeftPanel";
+import { setRightPanel } from "../../../../actions/setRightPanel";
+import { queryBuilder } from "../../helpers/querybuilder";
+import ShowLogsButton from "../Buttons/ShowLogsButton";
 import ValuesList from "./ValuesList";
+
+function panelAction(name, value) {
+    if (name === "left") {
+        return setLeftPanel(value);
+    }
+    return setRightPanel(value);
+}
 
 export default function ValuesSelector(props) {
     const { data, name } = props;
     const { dataSourceType, id } = data;
-    const panel = useSelector(store => store[name])
+    const dispatch = useDispatch();
+    const panelQuery = useSelector((store) => store[name]);
     const [labels, setLabels] = useState(props.labelsSelected);
 
     const labelsFiltered = useMemo(() => {
@@ -13,9 +25,7 @@ export default function ValuesSelector(props) {
             return labels.filter((f) => f.name !== "__name__");
         }
         return labels;
-
     }, [labels, dataSourceType]);
-
 
     const metricsSelection = useMemo(() => {
         if (dataSourceType === "metrics") {
@@ -24,6 +34,17 @@ export default function ValuesSelector(props) {
         return null;
     }, [labels, dataSourceType]);
 
+    const useQuery = () => {
+        const qu = queryBuilder(data.labels, data.expr);
+        const panel = [...panelQuery];
+        panel.forEach((query) => {
+            if (query.id === props.data.id) {
+                query.expr = qu;
+                query.labels = data.labels;
+            }
+        });
+        dispatch(panelAction(name, panel));
+    };
     useEffect(() => {
         setLabels(props.labelsSelected);
     }, [props.labelsSelected, setLabels]);
@@ -32,12 +53,29 @@ export default function ValuesSelector(props) {
         <div className="values-container">
             <div className="values-container-column">
                 {metricsSelection !== null && labels.length > 0 && (
-                    <ValuesList {...props} label={"__name__"} type={'metrics'} />
+                    <ValuesList
+                        {...props}
+                        label={"__name__"}
+                        type={"metrics"}
+                    />
                 )}
                 {labels &&
                     labelsFiltered?.map((label, key) => (
-                        <ValuesList {...props} label={label} key={key} type={'logs'} />
+                        <ValuesList
+                            {...props}
+                            label={label}
+                            key={key}
+                            type={"logs"}
+                        />
                     ))}
+            </div>
+            <div style={{ margin: "3px", marginButton: "6px" }}>
+                <ShowLogsButton
+                    onClick={useQuery}
+                    isMobile={false}
+                    alterText={"Use Query"}
+                    loading={false}
+                />
             </div>
         </div>
     );
